- Use new docker image for Python3 (same as KPI) - Use pip-tools for pip dependencies

+++ docker/init.bash
@@ -5,15 +5,16 @@ source /etc/profile
 
 echo 'KoBoCAT intializing...'
 
-oldpwd=$(pwd)
 cd "${KOBOCAT_SRC_DIR}"
 
-echo 'Synchronizing database.'
-python manage.py syncdb --noinput
+if [[ -z $DATABASE_URL ]]; then
+    echo "DATABASE_URL must be configured to run this server"
+    echo "example: 'DATABASE_URL=postgres://hostname:5432/dbname'"
+    exit 1
+fi
 
-echo 'Running migrations.'
+echo 'Running migrations...'
 python manage.py migrate --noinput
-
 
 rm -f /etc/cron.d/clean_up_tmp
 cp docker/cron/clean_up_tmp /etc/cron.d/
@@ -28,12 +29,18 @@ else
     echo "KoBoCAT media automatic backup schedule: ${KOBOCAT_MEDIA_BACKUP_SCHEDULE}"
 fi
 
+/bin/bash ${KOBOCAT_SRC_DIR}/docker/sync_static.sh
+
+# Keep it as is, not tested with Python3
 rm -rf /etc/profile.d/pydev_debugger.bash.sh
-if [[ -d /srv/pydev_orig && ! -z "${KOBOCAT_PATH_FROM_ECLIPSE_TO_PYTHON_PAIRS}" ]]; then
+if [[ -d /srv/pydev_orig && -n "${KOBOCAT_PATH_FROM_ECLIPSE_TO_PYTHON_PAIRS}" ]]; then
     echo 'Enabling PyDev remote debugging.'
     "${KOBOCAT_SRC_DIR}/docker/setup_pydev.bash"
 fi
 
+echo 'Cleaning up Celery PIDs...'
+rm -rf /tmp/celery*.pid
+
 echo 'KoBoCAT initialization complete.'
 
-cd $oldpwd
+exec /usr/bin/runsvdir /etc/service
